requirements.txt
fastapi==0.128.5
uvicorn==0.39.0
openai==2.17.0
pydantic==2.12.5
python-dotenv==1.2.1
python-docx==1.2.0
python-multipart==0.0.20
httpx==0.28.1
streamlit>=1.35.0
